Add border of vertices to top mesh
The points on the edge of the board are being cut off, so I'm extending the mesh with
a border of extra vertices.  In order to keep it simple, the boarder is spaced evenly
with the rest of the board, meaning the outermost tiles' texture is stretched to the
edge.  I think this is slightly better than it being cut off.