Move biome color theme definition to a separate instance
The color theme for biomes is now it's own class.  In addition to removing a lot of
clutter from the main BoardDisplay class, it also allows creating various themes
as prefab objects, so they can be swapped dynamically or set up per-scene.  Also,
since the BoardDisplay instance can reference a prefab theme not in the scene, you
are able to edit the theme while the game is running and have those changes
persist after the game stops.