Render a cube at the current cursor position
It took me a minute to understand how layer masks work here.  In order for this to
work, the GameBoard object (not it's children) should be on the 'Terrain' layer.
The Raycast method takes in a bitmask, so the int passed in needs to have the
bit corresponding to the layer's index set to 1.

TL;DR - you will need to add a 'Terrain' layer and assign the GameBoard to it for
mouse detection to work after this.